pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gpt-image-mcp"
version = "0.1.0"
description = "MCP server for image generation, editing, and analysis using OpenAI's gpt-image-1 model"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Du'An S. Lightfoot", email = "labeveryday@users.noreply.github.com"}
]
keywords = ["mcp", "image-generation", "openai", "gpt-image-1", "thumbnails", "youtube"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "mcp>=1.0.0",
    "fastmcp>=2.11.3",
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pillow>=10.0.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.5.0",
]

[project.optional-dependencies]
strands = [
    "strands-agents[openai]>=1.5.0",
    "strands-agents-tools>=0.2.4",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/labeveryday/gpt-image-mcp"
Repository = "https://github.com/labeveryday/gpt-image-mcp"
Issues = "https://github.com/labeveryday/gpt-image-mcp/issues"
Documentation = "https://github.com/labeveryday/gpt-image-mcp#readme"

[project.scripts]
gpt-image-mcp = "gpt_image_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/gpt_image_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "src/gpt_image_mcp",
    "tests",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
    "pyproject.toml",
]
exclude = [
    ".env",
    ".env.example",
    "img/",
    "examples/",
    "demo.py",
    "**/__pycache__",
    "**/.DS_Store",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]  # line-too-long: prompt strings are intentionally long

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.12.10",
]
